requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
numpy
ollama
PyPDF2
